Only chart model predictions, not raw data
Both provide unique value as output, but for this chart to be clear
and intuitive, it needs more x values -- and we can only add x
values to the chart of model predictions since we only have so much
raw data. Now that the charts are functioning properly they always
end up very similar anyway, so I don't think we lose much value by
narrowing down output to a single chart of model predictions.