Make contour size logic more direct and readable
I noted it was ugly from the beginning. It also ended up throwing
errors on Windows. This way does what we want in a much more
readable way. It may be slower, but the difference isn't
noticeable. End behavior seems to be the same (and if it weren't,
I might trust the new way more...).